Move 'generateDefaultIndexes' to utils
