+++ lib/index.js
@@ -1,7 +1,7 @@
-import { logSilenced } from './utils';
+import { logSilenced, generateDefaultIndexes } from './utils';
 import { Context } from './context';
 import { RenderLoop } from './render-loop';
-import { VertexSchema, generateDefaultIndexes } from './vertex-schema';
+import { VertexSchema } from './vertex-schema';
 import { VertexWriter } from './vertex-writer';
 
 export {
