+++ lib/utils.js
@@ -22,3 +22,11 @@ export function error(section, err) {
     isLogSilenced || console.error(`${section}: ${err.message}`);
     throw err;
 }
+
+export function generateDefaultIndexes(vertexCount) {
+    const data = new Array(vertexCount);
+    for (let i = 0; i < vertexCount; ++i) {
+        data[i] = i;
+    }
+    return data;
+}
