+++ lib/vertex-schema.js
@@ -72,11 +72,3 @@ export class VertexSchema {
         return this._isPacked;
     }
 }
-
-export function generateDefaultIndexes(vertexCount) {
-    const data = new Array(vertexCount);
-    for (let i = 0; i < vertexCount; ++i) {
-        data[i] = i;
-    }
-    return data;
-}
